Should always ask for least specific type you need
Only needs collection, so ask for collection. More or less the same
reason you don't put ArrayList in the parameter. You don't care what the
type is as long as it's a collection.

In consuming code I am trying to pass a map's keySet which is a Set.
Shouldn't have to convert it to an array first.